Move user code import to own module
